third_party/WebKit/LayoutTests/fast/viewport/viewport-legacy-merge-quirk-1-expected.txt

ALERT: viewport size 640x704 scale 1 with limits [1, 1] and userScalable false
